Drag fix: air drag is proportional to square the velocity squared
Previous implementation assumed it was directly proportional to the velocity.
The new behaviour is more accurate and enjoyable, but it could cause a lower acceleration curve, seeable mostly at very high speeds.
If this new behaviour would brake enjoyability unbalancing game times (cars too slow), we should reduce the drag coefficent, but keep the drag proportional to the square of the velocity. Anyway I don't think this is the case.



git-svn-id: svn://svn.code.sf.net/p/trigger-rally/code@871 6b6f3af4-9449-45d3-b87a-d38aa902195c