Storing offroad times individually for each vehicle instead of globally
The variables storing offroad times were stored globally in the TriggerGame class, and every vehicle shared them.
Now each vehicle has its own values, this is a needed fix to do if we want to implement any kind of multivehicle race (multyplayer, singleplayer with AI, mirror race).
Anyway displayed offroad time, and the time used for example to determine when the race finishes are taken by the user vehicle.
To know which vehicle is the user one, I created a PVehicle pointer in the TriggerGame class.



git-svn-id: svn://svn.code.sf.net/p/trigger-rally/code@873 6b6f3af4-9449-45d3-b87a-d38aa902195c