More accurate dust trails
Now dust trails are created when the wheel is touching the
ground (distance from it next to zero), making them more realistics.
Before the trail would be created anytime the wheel was also much more distant
from it. This was probably needed since in the past the 
computation of the contact point of the wheel with the
ground was less accurate, but that has been fixed with
commit [r983].



git-svn-id: svn://svn.code.sf.net/p/trigger-rally/code@973 6b6f3af4-9449-45d3-b87a-d38aa902195c