Now car dimensions are derived analizing the 3D model
Before they had to be specified in the .vehicle file.
In case autodetection doesn't work as the vehicle 
creator intended or for testing purpose, it's still possible
to specify them but it's not required anymore.

The dimensions are used by the game engine to compute the distribution
of the vehicle mass, so it affects stability, balance, ...

Having them automagically computed using the 3D model makes the game
feel more realistic and various.



git-svn-id: svn://svn.code.sf.net/p/trigger-rally/code@974 6b6f3af4-9449-45d3-b87a-d38aa902195c