Now the engine output will be distributed between the driving wheels
This seems obvious, but the previous implementation was different.
The output torque of the engine was directly wired to each wheel at the same way.
This meant for example that the sam engine would double is effective power when used
in AWD than when in FWD/RWD.
This messed up a couple of things around.

(And also violated thermodynamic)



git-svn-id: svn://svn.code.sf.net/p/trigger-rally/code@977 6b6f3af4-9449-45d3-b87a-d38aa902195c