Improved drag code and now drag is default to ON
The drag calculation algorythm has been improved and made more clever.
It alone determines what is the more realistic and accurate drag to apply,
and the values in the .vehicle file is just a coefficent to increase or
decrease the standard value.
The algorythm considers also the area of the section of the car.



git-svn-id: svn://svn.code.sf.net/p/trigger-rally/code@978 6b6f3af4-9449-45d3-b87a-d38aa902195c